If definitions has extra unused bytes, seek to right location
Previously, reading of definition levels stopped when there were enough to
satisfy the given row count for the data page in question. It turns out
there can be unused bytes after this - but that the total length of the
definitions block is correctly specified, so the extra bytes can simply
be skipped.