Allow encoding of object columns as bool/int
The only way to store int-with-None and bool-with-None in pandas
is in an object column. This adds object_encoding types 'bool' and
'int' (which can be auto-inferred) to store such columns.

Caveat: bool columns containing nulls get cast to float on load,
rather than creating expensive python obbjects. This is not then a
true round-trip, and this decision is open to discussion.