Add workaround for single-value-partition
In the case where data is partitioned, and the upmost partition
only has one possible value, we have no clear way to know that
it is meant as a partition rather than a location if passed a list
of paths.
As a workaround, provide extra parameter to ParquetFile (and merge)
to specify where the dataset root is.