COMPAT: for new pandas CategoricalDtype
Adds compatibility for the new pandas CategoricalDtype added in
https://github.com/pandas-dev/pandas/pull/16015

Most changes are around using `is_categorical_dtype` rather than str(dtype) ==
'category', as the repr of `Categorical.dtype` has changed to include the
categories and orderedness.

The only substantive change is in the update `catdef` step. Before we overwrite
the private `Categorical._categories` property, which no longer exists. Now we
call the (still private) `Categorical._set_categories` method. We need the
private method since pandas doesn't (and *probably* shouldn't) expose a way to
update a Categorical inplace with a different number of categories. This is OK
here, since all of our categories are "0" at the moment.