Use the python executable used to run the code (#484)
Currently `test_import_without_warning` uses the system default python executable, `python`.  This, however, may not be the same as the python executable that was used to run the test, and another version may not have fastparquet available.  This patch changes it to use the same executable used to run the test.